requirements/common.txt
psycopg2-binary==2.8.3
asgiref==3.2.7
Django==3.0.5
pytz==2019.3
sqlparse==0.3.1
django-modeltranslation==0.14.4
django-ckeditor==5.9.0
Pillow==7.1.1
djangorestframework==3.11.0
django-webpack-loader==0.7.0
django-filter==2.2.0
